Raster already distributed. Starting coding of the simulation itself.
Raster creation data are transferred from the Master node to the rest of them. The nodes themselves create those rasters, but only print/use the part they are in charge of. 

OPTIMIZATION for memory: delete not-belonging section of the rasters? Now, rasters are implemented with vector<vector<>>(N x M). For this optimization they should be Rectangles<>(N_i x M_i).