Added new feature to partition space evenly (pieces of ~equal area).
A new parameter in the configFile mpi::unevenPartitioning can be set to "true" to get a behaviour like before, or to "false" to split the K spaces of approximately equal area. In other words, the first mode splits by agents, the second splits by space.

Uneven partitioning (classic) is mainly useful when expecting uniformly distributed agents across the simulation space. As a rule of thumb, for dynamic creation of agents or agents around POIs/hot-spots, use the new even partitioning mode.